requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
google-genai==1.26.0
Pillow==11.3.0
python-telegram-bot==22.2
